COMP: fix warnings

+++ Modules/IGT/IGTToolManagement/mitkNavigationToolStorage.cpp
@@ -30,7 +30,7 @@ mitk::NavigationToolStorage::~NavigationToolStorage()
 
 bool mitk::NavigationToolStorage::DeleteTool(int number)
   {
-    if (number>m_ToolCollection.size()) return false;
+    if ((unsigned int)number > m_ToolCollection.size()) return false;
     std::vector<mitk::NavigationTool::Pointer>::iterator it = m_ToolCollection.begin() + number;
     m_ToolCollection.erase(it);
     return true;
